Update README.md
In this work we propose a new biophysical computational model of brain regions relevant to Parkinson's Disease (PD) based on local field potential data collected from the brain of marmoset monkeys. PD is a neurodegenerative disorder, linked to the death of dopaminergic neurons at the substantia nigra pars compacta, which affects the normal dynamics of the basal ganglia-thalamus-cortex (BG-T-C) neuronal circuit of the brain. Although there are multiple mechanisms underlying the disease, a complete description of those mechanisms and molecular pathogenesis are still missing, and there is still no cure. To address this gap, computational models that resemble neurobiological aspects found in animal models have been proposed. In our model, we performed a data-driven approach in which a set of biologically constrained parameters is optimised using differential evolution. Evolved models successfully resembled spectral signatures of local field potentials and single-neuron mean firing rates from healthy and parkinsonian marmoset brain data. {This is the first computational model of PD based on simultaneous electrophysiological recordings from seven brain regions of Marmoset monkeys.} Results indicate that the proposed model may facilitate the investigation of the mechanisms of PD and eventually support the development of new therapies. The DE method could also be applied to other computational neuroscience problems in which biological data is used to fit multi-scale models of brain circuits.